Add PE simplified agent startup script
This commit adds a new startup script to the available options. It
is based on the simplified agent (package-based) install available
with the latest version of Puppet Enterprise.
By default, it will work in collaboration with a PE Master of the
same operating system family. README update coming in a separate
commit.