Enable trusted autosigning for PE agents
This commit adds all the bits necessary to enable policy-based
autosigning of Puppet Enterprise agents in GCE with a PE master
also in GCE.

With changes in this commit, PE Masters classified using the
gce_compute::autosign class and assigned a compute-ro service scope
combined with PE Agents provisioned using the pe-simplified-agent
will connect to each other with no intervention on behalf of the user.

The startup script retrieves its unique id from GCE metadata and
stuffs it into the csr it makes to the PE master. The PE master
queries the google compute API for this metadata and compares it to
what's in the csr.